pofcc-375 - Moved angular specific code into the Service
